pyproject.toml
[project]
name = "genrec"
version = "0.1.0"
description = "Unified generative recommendation research framework"
authors = [
    { name = "Tiny_Snow", email = "futurelover10032@gmail.com" }
]
license = "MIT"
readme = "README.md"
keywords = ["recommendation", "generative-recommendation"]

[tool.poetry.dependencies]
python = ">=3.13,<3.15"

torch = { version = ">=2.5.1,<3.0", source = "pytorch-cu121" }
transformers = ">=4.57.3,<5.0"
datasets = ">=4.4.1,<5.0"
accelerate = ">=1.12.0,<2.0"
evaluate = ">=0.4.6,<1.0"
numpy = ">=2.3.5,<3.0"
scikit-learn = ">=1.7.2,<2.0"
rich = ">=14.2.0,<15.0"
tqdm = ">=4.67.1,<5.0"
pyyaml = ">=6.0.3,<7.0"
sentence-transformers = ">=5.1.2,<6.0"
jaxtyping = ">=0.3.3,<1.0"
numba = ">=0.63.1,<1.0"
tensorboardx = ">=2.6.4,<3.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0"
pytest-cov = ">=7.0.0,<8.0"
ruff = ">=0.14.8,<1.0"
pre-commit = ">=4.5.1,<5.0"
black = ">=25.12.0,<26.0"

[tool.black]
line-length = 120
target-version = ["py313"]
skip-string-normalization = true

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
